Approach to remove need to `spy()` in tests. This moves the client construction code into a separate class. The new class isn't covered at all, although most of the code in it wasn't covered before, either. Writing tests with `spy()` can be confusing, this change should help.
